added endpoint permissions

+++ app/web/routers/rooms.py
@@ -2,6 +2,7 @@ import logging
 from urllib.parse import quote, unquote
 
 from fastapi import APIRouter, Depends, Form, Request
+from starlette.datastructures import MutableHeaders
 from starlette.exceptions import HTTPException
 from starlette.responses import RedirectResponse
 
@@ -11,6 +12,9 @@ from app.store.database.queries.rooms import RoomDB
 from app.store.database.queries.users import UserDB
 from app.store.scheduler import operations as scheduler_operations
 from app.web.dependencies import get_current_user
+from app.web.permissions import (is_admin_is_room_own_or_member,
+                                 is_admin,
+                                 is_admin_is_room_own)
 from app.web.schemes import RemoveMemberConfirmation
 
 router = APIRouter()
@@ -19,31 +23,28 @@ logger = logging.getLogger(__name__)
 
 @router.get("/rooms", name='rooms')
 async def index(request: Request,
-                current_user: User = Depends(get_current_user)):
-    # TODO add permissions [superuser]
-    
-    rooms = await RoomDB.get_all_rooms()
-    
+                current_user: User = Depends(get_current_user),
+                permissions=Depends(is_admin), ):
     context = {
         'request': request,
         'current_user': current_user,
-        'rooms': rooms
+        'rooms': await RoomDB.get_all_rooms()
     }
     
     return templates.TemplateResponse(
-        'rooms//rooms.html', context=context)
+        'rooms/rooms.html', context=context)
 
 
 @router.get("/room/{room_number}", name='room')
 async def index(request: Request, room_number: int,
                 current_user: User = Depends(get_current_user),
-                ):
-    # TODO add permissions [superuser | owner | member]
-    # TODO добавить дату жеребьевки
-    
+                permissions=Depends(is_admin_is_room_own_or_member)):
     room = await RoomDB.get(room_number)
     if not room:
         raise HTTPException(status_code=404)
+   
+
+    
     members = await RoomDB.get_list_members(room_number=room_number)
     room_toss_time = scheduler_operations.get_task(room.number)
     if room_toss_time:
@@ -55,7 +56,6 @@ async def index(request: Request, room_number: int,
         'room_owner': await room.owner,
         'toss_time': room_toss_time,
         'users': members,
-        
     }
     
     return templates.TemplateResponse(
@@ -65,20 +65,28 @@ async def index(request: Request, room_number: int,
 @router.get("/room/{room_number}/remove_member/", name='user_leave_confirm')
 async def remove_member_conf(request: Request, room_number: int,
                              current_user: User = Depends(get_current_user),
-                             params=Depends(RemoveMemberConfirmation)
-                             ):
-    # TODO add permissions [superuser | owner if user is not owner of room]
-    # only superuser or owner
-    
+                             params=Depends(RemoveMemberConfirmation),
+                             permissions=Depends(
+                                 is_admin_is_room_own)):
+    """
+    The confirmation Endpoint to remove members from the room.
+    Permissions: Only the superuser and the owner of the room,
+                 but the owners cannot remove themselves from the room
+    """
+    is_room_owner = await RoomDB.is_owner(room_number=room_number,
+                                    user_id=params.dict().get('user_id'))
+    if is_room_owner:
+        raise HTTPException(status_code=403)
+        
     user_id = params.dict().get('user_id')
     context = {
         'request': request,
         'current_user': current_user,
         'number': room_number,
         'user': await UserDB.get_user_or_none(user_id),
-        'referer': quote(request.headers.get('referer'), safe='')
-        
+        'referer': quote(request.headers.get('referer'), safe=''),
     }
+    
     return templates.TemplateResponse(
         'users/leave_confirmation.html', context=context, status_code=200)
 
@@ -87,8 +95,20 @@ async def remove_member_conf(request: Request, room_number: int,
 async def remove_member(request: Request, user_id: int = Form(...),
                         room_number: int = Form(...),
                         referer: str = Form(...), confirm: bool = Form(...),
-                        current_user: User = Depends(get_current_user)):
-    # TODO add permissions [superuser | owner if user is not owner of room]
+                        current_user: User = Depends(get_current_user),
+                        ):
+    """
+    The endpoint to remove members from the room.
+    Permissions: Only the superuser and the owner of the room,
+                 but the owners cannot remove themselves from the room
+    """
+    
+    is_room_owner = await RoomDB.is_owner(room_number=room_number,
+                                          user_id=user_id)
+    logger.info(f'-----------{is_room_owner}')
+    if is_room_owner:
+        raise HTTPException(status_code=403)
+    logger.info(f'-----------{confirm}')
     if confirm:
         await RoomDB.remove_member(user_id, room_number)
